refactor: move G1 locomotion to robots/unitree_g1, update README and fix code smells
- Create src/roboharness/robots/unitree_g1/ package for robot-specific code,
  separating G1-specific GrootLocomotionController from the generic harness core
- Keep backwards-compatible re-export in controllers/locomotion.py
- Add conditional exports in controllers/__init__.py (WbcIkController)
- Update README: add LeRobot G1 section, Controllers table, Project Structure,
  improved G1 WBC description, expanded install options
- Fix version mismatch: __init__.py now matches pyproject.toml (0.1.1)
- Add docstring to core/__init__.py
- Update pyproject.toml coverage omit for new file paths
