+++ src/roboharness/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core framework components — Harness, Checkpoint, Capture, Controller protocol."""
