+++ src/roboharness/robots/__init__.py
@@ -1,0 +1,10 @@
+"""Robot-specific implementations.
+
+Each sub-package groups constants, controllers, and environment adapters for
+a specific robot platform. These are kept separate from the harness core so
+the framework stays robot-agnostic while still providing production-quality
+reference integrations.
+
+Available robots:
+  - ``unitree_g1``: Unitree G1 humanoid (29-DOF body, GR00T WBC policies)
+"""
