+++ src/roboharness/robots/unitree_g1/__init__.py
@@ -1,0 +1,31 @@
+"""Unitree G1 humanoid support — locomotion controllers and constants.
+
+Provides GR00T-based locomotion policies (Balance + Walk) and G1-specific
+joint configuration constants. Requires ``pip install roboharness[lerobot]``
+for ONNX runtime and HuggingFace model downloads.
+
+Usage::
+
+    from roboharness.robots.unitree_g1 import GrootLocomotionController
+
+    ctrl = GrootLocomotionController()
+    action = ctrl.compute(
+        command={"velocity": [0, 0, 0]},
+        state={"qpos": data.qpos, "qvel": data.qvel},
+    )
+"""
+
+from __future__ import annotations
+
+__all__: list[str] = []
+
+# Conditional exports — only available when [lerobot] extras are installed
+try:
+    from roboharness.robots.unitree_g1.locomotion import (
+        GrootLocomotionController,
+        HolosomaLocomotionController,
+    )
+
+    __all__ += ["GrootLocomotionController", "HolosomaLocomotionController"]
+except ImportError:
+    pass
